requirements.txt
discord.py>=2.3.2
yt_dlp
aiohttp
pynacl
spotipy
psycopg2-binary
sqlalchemy
cryptography
rapidfuzz
